refactor: group minio helpers in one file
